backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

# Image processing for EXIF extraction
Pillow==10.1.0

# Geospatial calculations
geopy==2.4.1

# Hashing and security
bcrypt==4.1.2
passlib==1.7.4

# Date and time utilities
python-dateutil==2.8.2

# UUID generation
uuid==1.30

# JSON schema validation (optional)
jsonschema==4.20.0
